+++ sansa-query/sansa-query-spark/src/test/scala/net/sansa_stack/query/spark/compliance/SPARQL11TestSuiteRunnerSpark.scala
@@ -1,14 +1,23 @@
 package net.sansa_stack.query.spark.compliance
 
 import com.holdenkarau.spark.testing.SharedSparkContext
+
 import net.sansa_stack.query.tests.SPARQLQueryEvaluationTestSuiteRunner
-import org.apache.jena.graph.NodeFactory
+import org.apache.jena.graph.{NodeFactory, Triple}
+import org.apache.jena.query.Query
+import org.apache.jena.rdf.model.Model
 import org.apache.jena.sparql.core.Var
 import org.apache.jena.sparql.engine.binding.{Binding, BindingFactory}
 import org.apache.jena.sparql.expr.NodeValue
+import org.apache.jena.sparql.resultset.SPARQLResult
+import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 import org.scalatest.{ConfigMap, Suite}
+
+import net.sansa_stack.query.spark.api.domain.{QueryEngineFactory, QueryExecutionFactorySpark}
+
+import scala.collection.JavaConverters._
 
 /**
  * SPARQL 1.1 test suite runner on Apache Spark.
@@ -36,13 +45,6 @@ abstract class SPARQL11TestSuiteRunnerSpark
       .set("spark.kryo.registrator", "net.sansa_stack.rdf.spark.io.JenaKryoRegistrator")
   ).getOrCreate()
 
-//  override def beforeAll(): Unit = {
-////    super.beforeAll()
-//    conf.set("spark.sql.crossJoin.enabled", "true")
-//      .set("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-//      .set("spark.kryo.registrator", "net.sansa_stack.rdf.spark.io.JenaKryoRegistrator")
-//    _spark = SparkSession.builder.config(conf).master("local[1]").getOrCreate()
-//  }
 
   override val invokeBeforeAllAndAfterAllEvenIfNoTestsAreExpected: Boolean = false
 
@@ -62,6 +64,58 @@ abstract class SPARQL11TestSuiteRunnerSpark
     super.afterAll()
     spark.stop()
     _spark = null
+  }
+
+  def getEngineFactory: QueryEngineFactory
+
+  lazy val engineFactory: QueryEngineFactory = getEngineFactory
+
+  val db = "TEST"
+
+  var previousModel: Model = _
+  var triplesRDD: RDD[Triple] = _
+  var qef: QueryExecutionFactorySpark = _
+
+  override def runQuery(query: Query, data: Model): SPARQLResult = {
+    // do some caching here to avoid reloading the same data
+    if (data != previousModel) {
+      // we drop the Spark database to remove all tables
+      spark.sql(s"DROP DATABASE IF EXISTS $db")
+
+      // distribute on Spark
+      triplesRDD = spark.sparkContext.parallelize(data.getGraph.find().toList.asScala)
+
+      // we create a Spark database here to keep the implicit partitioning separate
+
+      spark.sql(s"CREATE DATABASE IF NOT EXISTS $db")
+      spark.sql(s"USE $db")
+
+      qef = engineFactory.create(triplesRDD)
+
+      previousModel = data
+    }
+
+    val qe = qef.createQueryExecution(query)
+
+    // produce result based on query type
+    val result = if (query.isSelectType) { // SELECT
+      val rs = qe.execSelect()
+      new SPARQLResult(rs)
+    } else if (query.isAskType) { // ASK
+      val b = qe.execAsk()
+      new SPARQLResult(b)
+    } else if (query.isConstructType) { // CONSTRUCT
+      val triples = qe.execConstruct()
+      new SPARQLResult(triples)
+    } else { // DESCRIBE todo
+      fail("unsupported query type: DESCRIBE")
+      null
+    }
+    // clean up
+
+    qe.close()
+
+    result
   }
