+++ sansa-rdf/sansa-rdf-common/src/main/scala/net/sansa_stack/rdf/common/partition/r2rml/R2rmlUtils.scala
@@ -248,7 +248,7 @@ object R2rmlUtils {
    * @param model the model
    * @return
    */
-  def triplesMapsForPredicate(predicate : Property, model: Model): Iterator[TriplesMap] = {
+  def triplesMapsForPredicate(predicate: Property, model: Model): Iterator[TriplesMap] = {
     model
       .listResourcesWithProperty(RR.subjectMap).asScala
       .map(_.as(classOf[TriplesMap]))
